extract ExternalIps component that doesn't know about primary
