+++ app/pages/system/SiloPage.tsx
@@ -8,7 +8,14 @@
 import { Link, Outlet, type LoaderFunctionArgs } from 'react-router-dom'
 
 import { apiQueryClient, usePrefetchedApiQuery } from '@oxide/api'
-import { DateCell, DefaultCell, linkCell, TruncateCell, useQueryTable } from '@oxide/table'
+import {
+  DateCell,
+  DefaultCell,
+  EmptyCell,
+  linkCell,
+  TruncateCell,
+  useQueryTable,
+} from '@oxide/table'
 import {
   Badge,
   buttonStyle,
@@ -26,8 +33,6 @@ import {
 
 import { getSiloSelector, useSiloSelector } from 'app/hooks'
 import { pb } from 'app/util/path-builder'
-
-import { EmptyCellContent } from '../project/instances/instance/tabs/NetworkingTab'
 
 const EmptyState = () => (
   <EmptyMessage icon={<Cloud16Icon />} title="No identity providers" />
@@ -75,7 +80,7 @@ export function SiloPage() {
         Fleet role mapping <RoleMappingTooltip />
       </h2>
       {roleMapPairs.length === 0 ? (
-        <EmptyCellContent />
+        <EmptyCell />
       ) : (
         <ul className="space-y-3">
           {roleMapPairs.map(([siloRole, fleetRole]) => (
