+++ libs/table/cells/index.ts
@@ -11,6 +11,7 @@ export * from './Cell'
 export * from './DateCell'
 export * from './DefaultCell'
 export * from './EnabledCell'
+export * from './EmptyCell'
 export * from './FirewallFilterCell'
 export * from './InstanceResourceCell'
 export * from './InstanceStatusCell'
